% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunnettT3Test.R
\name{dunnettT3Test}
\alias{dunnettT3Test}
\alias{dunnettT3Test.default}
\alias{dunnettT3Test.formula}
\alias{dunnettT3Test.aov}
\title{Dunnett's T3 Test}
\usage{
dunnettT3Test(x, ...)

\method{dunnettT3Test}{default}(x, g, ...)

\method{dunnettT3Test}{formula}(formula, data, subset, na.action, ...)

\method{dunnettT3Test}{aov}(x, ...)
}
\arguments{
\item{x}{a numeric vector of data values, a list of numeric data
vectors or a fitted model object, usually an \link[stats]{aov} fit.}

\item{\dots}{further arguments to be passed to or from methods.}

\item{g}{a vector or factor object giving the group for the
corresponding elements of \code{"x"}.
Ignored with a warning if \code{"x"} is a list.}

\item{formula}{a formula of the form \code{response ~ group} where
\code{response} gives the data values and \code{group} a vector or
factor of the corresponding groups.}

\item{data}{an optional matrix or data frame (or similar: see
\code{\link{model.frame}}) containing the variables in the
formula \code{formula}.  By default the variables are taken from
\code{environment(formula)}.}

\item{subset}{an optional vector specifying a 
subset of observations to be used.}

\item{na.action}{a function which indicates what should happen when
the data contain \code{NA}s.  Defaults to \code{getOption("na.action")}.}
}
\value{
A list with class \code{"PMCMR"} containing the following components:
\describe{
 \item{method}{a character string indicating what type of test was performed.}
 \item{data.name}{a character string giving the name(s) of the data.}
 \item{statistic}{lower-triangle matrix of the estimated
quantiles of the pairwise test statistics.}
 \item{p.value}{lower-triangle matrix of the p-values for the pairwise tests.}
 \item{alternative}{a character string describing the alternative hypothesis.}
\item{p.adjust.method}{a character string describing the method for p-value
adjustment.}
\item{model}{a data frame of the input data.}
\item{dist}{a string that denotes the test distribution.}
}
}
\description{
Performs Dunnett's all-pairs comparison test for normally distributed
data with unequal variances.
}
\details{
For all-pairs comparisons in an one-factorial layout
with normally distributed residuals but unequal groups variances
the T3 test of Dunnett can be performed.
Let \eqn{X_{ij}} denote a continuous random variable
with the \eqn{j}-the realization (\eqn{1 \le j \le n_i})
in the \eqn{i}-th group (\eqn{1 \le i \le k}). Furthermore, the total
sample size is \eqn{N = \sum_{i=1}^k n_i}. A total of \eqn{m = k(k-1)/2}
hypotheses can be tested: The null hypothesis is
H\eqn{_{ij}: \mu_i = \mu_j ~~ (i \ne j)} is tested against the alternative
A\eqn{_{ij}: \mu_i \ne \mu_j} (two-tailed). Dunnett T3 all-pairs
test statistics are given by

\deqn{
 t_{ij} \frac{\bar{X}_i - \bar{X_j}}
 {\left( s^2_j / n_j + s^2_i / n_i \right)^{1/2}}, ~~
 (i \ne j)
}{%
 SEE PDF
}

with \eqn{s^2_i} the variance of the \eqn{i}-th group.
The null hypothesis is rejected (two-tailed) if

\deqn{
 \mathrm{Pr} \left\{ |t_{ij}| \ge T_{v_{ij}\rho_{ij}\alpha'/2} | \mathrm{H} \right\}_{ij} =
 \alpha,
}{%
 SEE PDF
}

with Welch's approximate solution for calculating the degree of freedom.

\deqn{
 v_{ij} = \frac{\left( s^2_i / n_i + s^2_j / n_j \right)^2}
 {s^4_i / n^2_i \left(n_i - 1\right) + s^4_j / n^2_j \left(n_j - 1\right)}.
}{%
 SEE PDF
}

The \eqn{p}-values are computed from the
studentized maximum modulus distribution
that is the equivalent of the multivariate t distribution
with \eqn{\rho_{ii} = 1, ~ \rho_{ij} = 0 ~ (i \ne j)}.
The function \code{\link[mvtnorm]{pmvt}} is used to
calculate the \eqn{p}-values.
}
\examples{
fit <- aov(weight ~ feed, chickwts)
shapiro.test(residuals(fit))
bartlett.test(weight ~ feed, chickwts)
anova(fit)

## also works with fitted objects of class aov
res <- dunnettT3Test(fit)
summary(res)
summaryGroup(res)

}
\references{
C. W. Dunnett (1980) Pair wise multiple comparisons in the unequal
 variance case, \emph{Journal of the American Statistical
 Association} \bold{75}, 796--800.
}
\seealso{
\code{\link[mvtnorm]{pmvt}}
}
\concept{parametric}
\keyword{htest}
