% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jonckheere.test.R
\name{jonckheereTest}
\alias{jonckheereTest}
\alias{jonckheereTest.default}
\alias{jonckheereTest.formula}
\title{Testing against Ordered Alternatives (Jonckheere-Terpstra Test)}
\usage{
jonckheereTest(x, ...)

\method{jonckheereTest}{default}(
  x,
  g,
  alternative = c("two.sided", "greater", "less"),
  continuity = FALSE,
  ...
)

\method{jonckheereTest}{formula}(
  formula,
  data,
  subset,
  na.action,
  alternative = c("two.sided", "greater", "less"),
  continuity = FALSE,
  ...
)
}
\arguments{
\item{x}{a numeric vector of data values, or a list of numeric data
vectors.}

\item{\dots}{further arguments to be passed to or from methods.}

\item{g}{a vector or factor object giving the group for the
corresponding elements of \code{"x"}.
Ignored with a warning if \code{"x"} is a list.}

\item{alternative}{the alternative hypothesis. Defaults to \code{"two.sided"}.}

\item{continuity}{logical indicator whether a continuity correction
shall be performed. Defaults to \code{FALSE}.}

\item{formula}{a formula of the form \code{response ~ group} where
\code{response} gives the data values and \code{group} a vector or
factor of the corresponding groups.}

\item{data}{an optional matrix or data frame (or similar: see
\code{\link{model.frame}}) containing the variables in the
formula \code{formula}.  By default the variables are taken from
\code{environment(formula)}.}

\item{subset}{an optional vector specifying a 
subset of observations to be used.}

\item{na.action}{a function which indicates what should happen when
the data contain \code{NA}s.  Defaults to \code{getOption("na.action")}.}
}
\value{
A list with class \code{"htest"} containing the following components:
\describe{
 \item{method}{a character string indicating what type of test was performed.}
 \item{data.name}{a character string giving the name(s) of the data.}
 \item{statistic}{the estimated quantile of the test statistic.}
 \item{p.value}{the p-value for the test.}
 \item{parameter}{the parameters of the test statistic, if any.}
 \item{alternative}{a character string describing the alternative hypothesis.}
\item{estimates}{the estimates, if any.}
\item{null.value}{the estimate under the null hypothesis, if any.}
}
}
\description{
Performs the Jonckheere-Terpstra test for testing against ordered alternatives.
}
\details{
The null hypothesis, H\eqn{_0: \theta_1 = \theta_2 = \ldots = \theta_k}
is tested against a simple order hypothesis,
H\eqn{_\mathrm{A}: \theta_1 \le \theta_2 \le \ldots \le
\theta_k,~\theta_1 < \theta_k}.

The p-values are estimated from the standard normal distribution.
}
\note{
\code{jonckheereTest(x, g, alternative = "two.sided", continuity = TRUE)} is
equivalent to

\code{cor.test(x, as.numeric(g), method = "kendall", alternative = "two.sided", continuity = TRUE)}
}
\section{Source}{

The code for the computation of the standard deviation
for the Jonckheere-Terpstra test in the presence of ties was taken from:\cr

Kloke, J., McKean, J. (2016)
\CRANpkg{npsm}: Package for Nonparametric Statistical Methods using R.
R package version 0.5. \url{https://CRAN.R-project.org/package=npsm}
}

\examples{
## Example from Sachs (1997, p. 402)
x <- c(106, 114, 116, 127, 145,
       110, 125, 143, 148, 151,
       136, 139, 149, 160, 174)
g <- gl(3,5)
levels(g) <- c("A", "B", "C")

## Chacko's test
chackoTest(x, g)

## Cuzick's test
cuzickTest(x, g)

## Johnson-Mehrotra test
johnsonTest(x, g)

## Jonckheere-Terpstra test
jonckheereTest(x, g)

## Le's test
leTest(x, g)

## Spearman type test
spearmanTest(x, g)

## Murakami's BWS trend test
bwsTrendTest(x, g)

}
\references{
Jonckheere, A. R. (1954) A distribution-free k-sample test
against ordered alternatives. \emph{Biometrica} \bold{41}, 133–145.

Kloke, J., McKean, J. W. (2015) \emph{Nonparametric statistical methods using R}.
Boca Raton, FL: Chapman & Hall/CRC.
}
\seealso{
\code{\link{kruskalTest}} and \code{\link{shirleyWilliamsTest}}
of the package \pkg{PMCMRplus},
\code{\link[stats]{kruskal.test}} of the library \pkg{stats}.
}
\concept{ordered-alternatives}
\concept{trend-test}
\keyword{htest}
\keyword{nonparametric}
