% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pentosan.R
\docType{data}
\name{Pentosan}
\alias{Pentosan}
\title{Pentosan Dataset}
\format{A data frame with 189 obs. of  3 variables:
\describe{
 \item{value}{numeric, test result (no unit specified)}
 \item{lab}{factor, identifier of the lab (1--7)}
 \item{material}{factor, identifier of the level of the material (A--I)}
}}
\source{
Tab. 8, Practice E 691, 2005, \emph{Standard Practice for
Conducting an Interlaboratory Study to Determine the
Precision of a Test Method}, ASTM International.
}
\description{
A benchmark dataset of an interlaboratory study for
determining the precision of a test method
on several levels of the material Pentosan.
}
\keyword{datasets}
