% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mandel.R
\name{plot.mandel}
\alias{plot.mandel}
\title{Plotting mandel Objects}
\usage{
\method{plot}{mandel}(x, alpha = 0.005, ...)
}
\arguments{
\item{x}{an object with class \code{"mandel"}.}

\item{alpha}{level of significance. Defaults to \code{0.005}.}

\item{\ldots}{further arguments, currently ignored.}
}
\description{
Plotting method for objects inheriting from class \code{"mandel"}.
}
\examples{
## 
\dontrun{
data(Pentosan)
md <- mandelkTest(value ~ lab, Pentosan, subset = (material == "B"))
plot(md)
}
}
\seealso{
\code{demo(Pentosan)}
}
