% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesdTest.R
\name{print.gesdTest}
\alias{print.gesdTest}
\title{gesdTest Printing}
\usage{
\method{print}{gesdTest}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{\ldots}{further arguments. Currently ignored.}
}
\description{
\code{print.gesdTest} is the \emph{gesdTest} method of the generic
\code{\link{print}} function which prints its argument
and returns it \emph{invisibly} (via \code{\link{invisible}(x)}).
}
\keyword{print}
