% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mandel.R
\name{print.mandel}
\alias{print.mandel}
\alias{print}
\title{Mandel Printing}
\usage{
\method{print}{mandel}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{\ldots}{further arguments. Currently ignored.}
}
\description{
\code{print.mandel} is the \emph{mandel} method of the generic
\code{\link{print}} function which prints its argument
and returns it \emph{invisibly} (via \code{\link{invisible}(x)}).
}
\seealso{
\code{\link{mandelhTest}}, \code{\link{mandelkTest}}
}
\keyword{print}
