% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.osrt.R
\name{print.osrt}
\alias{print.osrt}
\title{osrt Printing}
\usage{
\method{print}{osrt}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{\ldots}{further arguments. Currently ignored.}
}
\description{
\code{print.osrt} is the \emph{osrt} method of the generic
\code{\link{print}} function which prints its argument
and returns it \emph{invisibly} (via \code{\link{invisible}(x)}).
}
\seealso{
summary.osrt
}
\keyword{print}
