% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PMCMR.R
\name{print.PMCMR}
\alias{print.PMCMR}
\title{PMCMR Printing}
\usage{
\method{print}{PMCMR}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{\ldots}{further arguments. Currently ignored.}
}
\description{
\code{print.PMCMR} is the \emph{PMCMR} method of the generic
\code{\link{print}} function which prints its argument
and returns it \emph{invisibly} (via \code{\link{invisible}(x)}).
}
\keyword{print}
