% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.powerOneWayPMCMR.R
\name{print.powerOneWayPMCMR}
\alias{print.powerOneWayPMCMR}
\title{PowerOneWayPMCMR Printing}
\usage{
\method{print}{powerOneWayPMCMR}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{\ldots}{further arguments. Currently ignored.}
}
\description{
\code{print.powerOneWayPMCMR} is the
\emph{powerOneWayPMCMR} method of the generic
\code{\link{print}} function which prints its argument
and returns it \emph{invisibly} (via \code{\link{invisible}(x)}).
}
\keyword{print}
