% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.powerPMCMR.R
\name{print.powerPMCMR}
\alias{print.powerPMCMR}
\title{PowerPMCMR Printing}
\usage{
\method{print}{powerPMCMR}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{\ldots}{further arguments. Currently ignored.}
}
\description{
\code{print.powerPMCMR} is the \emph{powerPMCMR} method of the generic
\code{\link{print}} function which prints its argument
and returns it \emph{invisibly} (via \code{\link{invisible}(x)}).
}
\seealso{
\code{\link{powerMCTests}}, \code{\link{powerOneWayTests}}
}
\keyword{print}
