% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.steelTest.R
\name{print.steel}
\alias{print.steel}
\title{steel Printing}
\usage{
\method{print}{steel}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{\ldots}{further arguments. Currently ignored.}
}
\description{
\code{print.steel} is the \emph{steel} method of the generic
\code{\link{print}} function which prints its argument
and returns it \emph{invisibly} (via \code{\link{invisible}(x)}).
}
\keyword{print}
