% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.trendPMCMR.R
\name{print.trendPMCMR}
\alias{print.trendPMCMR}
\title{trendPMCMR Printing}
\usage{
\method{print}{trendPMCMR}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{\ldots}{further arguments. Currently ignored.}
}
\description{
\code{print.trendPMCMR} is the \emph{trendPMCMR} method of the generic
\code{\link{print}} function which prints its argument
and returns it \emph{invisibly} (via \code{\link{invisible}(x)}).
}
\keyword{print}
