% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.williamsTest.R
\name{print.williams}
\alias{print.williams}
\title{williams Printing}
\usage{
\method{print}{williams}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{\ldots}{further arguments. Currently ignored.}
}
\description{
\code{print.williams} is the \emph{williams} method of the generic
\code{\link{print}} function which prints its argument
and returns it \emph{invisibly} (via \code{\link{invisible}(x)}).
}
\keyword{print}
