% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shanTest.R
\name{shanTest}
\alias{shanTest}
\alias{shanTest.default}
\alias{shanTest.formula}
\title{Testing against Ordered Alternatives (Shan-Young-Kang Test)}
\usage{
shanTest(x, ...)

\method{shanTest}{default}(x, g, alternative = c("greater", "less"), ...)

\method{shanTest}{formula}(
  formula,
  data,
  subset,
  na.action,
  alternative = c("greater", "less"),
  ...
)
}
\arguments{
\item{x}{a numeric vector of data values, or a list of numeric data
vectors.}

\item{\dots}{further arguments to be passed to or from methods.}

\item{g}{a vector or factor object giving the group for the
corresponding elements of \code{"x"}.
Ignored with a warning if \code{"x"} is a list.}

\item{alternative}{the alternative hypothesis.
Defaults to \code{"greater"}.}

\item{formula}{a formula of the form \code{response ~ group} where
\code{response} gives the data values and \code{group} a vector or
factor of the corresponding groups.}

\item{data}{an optional matrix or data frame (or similar: see
\code{\link{model.frame}}) containing the variables in the
formula \code{formula}.  By default the variables are taken from
\code{environment(formula)}.}

\item{subset}{an optional vector specifying a 
subset of observations to be used.}

\item{na.action}{a function which indicates what should happen when
the data contain \code{NA}s.  Defaults to \code{getOption("na.action")}.}
}
\value{
A list with class \code{"htest"} containing the following components:
\describe{
 \item{method}{a character string indicating what type of test was performed.}
 \item{data.name}{a character string giving the name(s) of the data.}
 \item{statistic}{the estimated quantile of the test statistic.}
 \item{p.value}{the p-value for the test.}
 \item{parameter}{the parameters of the test statistic, if any.}
 \item{alternative}{a character string describing the alternative hypothesis.}
\item{estimates}{the estimates, if any.}
\item{null.value}{the estimate under the null hypothesis, if any.}
}
}
\description{
Performs the Shan-Young-Kang test for testing against ordered alternatives.
}
\details{
The null hypothesis, H\eqn{_0: \theta_1 = \theta_2 = \ldots = \theta_k}
is tested against a simple order hypothesis,
H\eqn{_\mathrm{A}: \theta_1 \le \theta_2 \le \ldots \le
\theta_k,~\theta_1 < \theta_k}.

Let \eqn{R_{ij}} be the rank of \eqn{X_{ij}},
where \eqn{X_{ij}} is jointly ranked
from \eqn{\left\{1, 2, \ldots, N \right\}, ~~ N = \sum_{i=1}^k n_i},
the the test statistic is

\deqn{
 S = \sum_{i = 1}^{k-1} \sum_{j = i + 1}^k D_{ij},
}{%
 SEE PDF
}

with
\deqn{
 D_{ij} = \sum_{l = 1}^{n_i} \sum_{m=1}^{n_j}  \left(R_{jm} - R_{il} \right)~ \mathrm{I}\left(X_{jm} > X_{il} \right),
}{%
 SEE PDF
}

where

\deqn{
 \mathrm{I}(u) = \left\{
 \begin{array}{c}
     1, \qquad \forall~ u > 0 \\
     0, \qquad \forall~ u \le 0
  \end{array}
  \right.
.}{%
 SEE PDF.
}

The test statistic is asymptotically normal distributed:
\deqn{
 z = \frac{S - \mu_{\mathrm{S}}}{\sqrt{s^2_{\mathrm{S}}}}
}{
 SEE PDF
}

The p-values are estimated from the standard normal distribution.
}
\note{
The variance estimation (see Theorem 2.1, Shan et al. 2014)
can become negative for certain combinations of \eqn{N,~n_i,~k
\qquad (1 \le i \le k)}. In these cases the function will return
a warning and the returned p-value will be \code{NaN}.
}
\examples{
## Example from Sachs (1997, p. 402)
x <- c(106, 114, 116, 127, 145,
       110, 125, 143, 148, 151,
       136, 139, 149, 160, 174)
g <- gl(3,5)
levels(g) <- c("A", "B", "C")

## Chacko's test
chackoTest(x, g)

## Cuzick's test
cuzickTest(x, g)

## Johnson-Mehrotra test
johnsonTest(x, g)

## Jonckheere-Terpstra test
jonckheereTest(x, g)

## Le's test
leTest(x, g)

## Spearman type test
spearmanTest(x, g)

## Murakami's BWS trend test
bwsTrendTest(x, g)

## Fligner-Wolfe test
flignerWolfeTest(x, g)

## Shan-Young-Kang test
shanTest(x, g)

}
\references{
Shan, G., Young, D., Kang, L. (2014) A New Powerful Nonparametric
Rank Test for Ordered Alternative Problem. PLOS ONE 9, e112924.
https://doi.org/10.1371/journal.pone.0112924
}
\seealso{
\code{\link{kruskalTest}} and \code{\link{shirleyWilliamsTest}}
of the package \pkg{PMCMRplus},
\code{\link[stats]{kruskal.test}} of the library \pkg{stats}.
}
\concept{trendtest}
\keyword{htest}
\keyword{nonparametric}
