% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesdTest.R
\name{summary.gesdTest}
\alias{summary.gesdTest}
\title{Summarize an gesdTest Object}
\usage{
\method{summary}{gesdTest}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"gesdTest"}.}

\item{\dots}{further arguments. Currenly ignored.}
}
\description{
Summarize an object of class \emph{gesdTest}.
}
\keyword{methods}
