% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mandel.R
\name{summary.mandel}
\alias{summary.mandel}
\title{Object Summary for class \code{"mandel"}}
\usage{
\method{summary}{mandel}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"mandel"} for
which a summary is desired.}

\item{\ldots}{further arguments. Currently ignored.}
}
\description{
\code{summary.mandel} is a function
used to produce result summaries of the results of
the functions \code{\link{mandelhTest}} or \code{\link{mandelkTest}}.
}
\seealso{
\code{\link{mandelhTest}}, \code{\link{mandelkTest}}
}
\keyword{methods}
