% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.osrt.R
\name{summary.osrt}
\alias{summary.osrt}
\title{Summarize an osrt Object}
\usage{
\method{summary}{osrt}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"osrt"}.}

\item{\dots}{further arguments. Currenly ignored.}
}
\description{
Summarize an object of class \emph{osrt}.
}
\seealso{
\code{\link{print.osrt}}.
}
\keyword{methods}
