% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PMCMR.R
\name{summary.PMCMR}
\alias{summary.PMCMR}
\title{Summarize an PMCMR Object}
\usage{
\method{summary}{PMCMR}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"PMCMR"}.}

\item{\dots}{further arguments. Currenly ignored.}
}
\value{
A detailed output of all pairwise hypotheses,
the test statistics, the corresponding p-values and
symbols that indicates the level of significance.
}
\description{
Summarize an object of class \emph{PMCMR}.
}
\examples{
ans <- vanWaerdenAllPairsTest(count ~ spray, InsectSprays)
summary(ans)
}
\seealso{
\code{\link{print.PMCMR}}, \code{\link{summaryGroup}}.
}
\keyword{models}
