% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.steelTest.R
\name{summary.steel}
\alias{summary.steel}
\title{Summarize a steel Object}
\usage{
\method{summary}{steel}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"steel"}.}

\item{\dots}{further arguments. Currenly ignored.}
}
\value{
A detailed output of all pairwise hypotheses,
the test statistics, the corresponding p-values and
symbols that indicates the level of significance.
}
\description{
Summarize an object of class \emph{steel}.
}
\examples{
ans <- vanWaerdenAllPairsTest(count ~ spray, InsectSprays)
summary(ans)
}
\seealso{
\code{\link{print.steel}}, \code{\link{summaryGroup}}.
}
\keyword{methods}
