% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.trendPMCMR.R
\name{summary.trendPMCMR}
\alias{summary.trendPMCMR}
\title{Summarize an trendPMCMR Object}
\usage{
\method{summary}{trendPMCMR}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"trendPMCMR"}.}

\item{\dots}{further arguments. Currenly ignored.}
}
\value{
A detailed output of all pairwise hypotheses,
the test statistics, the corresponding p-values and
symbols that indicates the level of significance.
}
\description{
Summarize an object of class \emph{trendPMCMR}.
}
\seealso{
\code{\link{print.trendPMCMR}}
}
\keyword{methods}
