% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.williamsTest.R
\name{summary.williams}
\alias{summary.williams}
\title{Summarize an williams Object}
\usage{
\method{summary}{williams}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"williams"}.}

\item{\dots}{further arguments. Currenly ignored.}
}
\value{
A detailed output of all pairwise hypotheses,
the test statistics, the corresponding p-values and
symbols that indicates the level of significance.
}
\description{
Summarize an object of class \emph{williams}.
}
\examples{
ans <- vanWaerdenAllPairsTest(count ~ spray, InsectSprays)
summary(ans)
}
\seealso{
\code{\link{print.williams}}, \code{\link{summaryGroup}}.
}
\keyword{models}
