% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PMCMR.R
\name{summaryGroup}
\alias{summaryGroup}
\title{Grouped Summary of an PMCMR Object}
\usage{
summaryGroup(x, alpha = 0.05, ...)
}
\arguments{
\item{x}{an object of class \code{"PMCMR"}.}

\item{alpha}{the selected alpha-level. Defaults to 0.05.}

\item{\dots}{further arguments. Currently ignored.}
}
\value{
Provides summary statistics for each factor level
and a letter symbol, whereas different letters indicate
significant differences between factor levels based on the
selected level of alpha.
}
\description{
Performes a grouped summary on an PMCMR object.
}
\seealso{
\code{\link{summary.PMCMR}}
}
\keyword{methods}
