% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toTidy.R
\name{toTidy}
\alias{toTidy}
\title{Convert a PMCMR or osrt Object to a Data.Frame}
\usage{
toTidy(mod, ...)
}
\arguments{
\item{mod}{an object of class \code{"PMCMR"}, \code{"trendPMCMR"} or \code{"osrt"}.}

\item{\ldots}{further arguments. Currently ignored.}
}
\value{
A data.frame.
}
\description{
The functions converts a list object of class \code{"PMCMR"}
or \code{"osrt"} into a data.frame.
}
\examples{
res <- tukeyTest(weight ~ Diet, data = ChickWeight, subset = Time == 21)
toTidy(res)

}
\author{
Indrajeet Patil (via email, 2020-1022),
modified by Thorsten Pohlert
}
\keyword{utilities}
